feat: user free space check

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('space')->default(0);
+            $table->decimal('space')->default(0);
             $table->string('subscription_type')->default('Free');
             $table->rememberToken();
             $table->timestamps();
